+++ projects/cct/src/element/element_button.cpp
@@ -117,5 +117,5 @@ void element_button::handle_event(SDL_Event *event, sdl_helper *helper)
 element_button *element_button::read_from_json(const json &j, SDL_Point *default_dim)
 {
 	return new element_button(j[CFG_ID].string_value(), read_position(j), read_mapping(j, default_dim),
-							  j[CFG_KEY_CODE].number_value(), j[CFG_Z_LEVEL].number_value());
+							  uint16_t(j[CFG_KEY_CODE].number_value()), uint8_t(j[CFG_Z_LEVEL].number_value()));
 }
