+++ projects/cct/src/element/element_texture.cpp
@@ -79,5 +79,5 @@ void element_texture::update_settings(dialog_element_settings *dialog)
 element_texture *element_texture::read_from_json(const json &j, SDL_Point *default_dim)
 {
 	return new element_texture(j[CFG_ID].string_value(), read_position(j), read_mapping(j, default_dim),
-							   j[CFG_Z_LEVEL].number_value());
+							   uint8_t(j[CFG_Z_LEVEL].number_value()));
 }
